+++ .devcontainer/devcontainer.json
@@ -10,9 +10,9 @@
 			// Use the -bullseye variants on local arm64/Apple Silicon.
 			"VARIANT": "17-bullseye",
 			// Options
-			"INSTALL_MAVEN": "true",
+			"INSTALL_MAVEN": "false",
 			"INSTALL_GRADLE": "true",
-			"NODE_VERSION": "none"
+			"NODE_VERSION": "lts/*"
 		}
 	},
 
@@ -26,10 +26,13 @@
 			
 			// Add the IDs of extensions you want installed when the container is created.
 			"extensions": [
-				"vscjava.vscode-java-pack"
+				"vscjava.vscode-java-pack",
+				"vscjava.vscode-gradle"
 			]
 		}
 	},
+
+	"postCreateCommand": "gh auth login",
 
 	// Use 'forwardPorts' to make a list of ports inside the container available locally.
 	// "forwardPorts": [],
@@ -39,17 +42,8 @@
 
 	// Comment out to connect as root instead. More info: https://aka.ms/vscode-remote/containers/non-root.
 	"remoteUser": "vscode",
-
-	"features": { 
-		"ghcr.io/devcontainers/features/github-cli": {},
-		"ghcr.io/devcontainers/features/python:1": {
-            		"version": "latest"
-        		}
-	},
-
-	"onCreateCommand": "./scripts/onCreateCommand.sh",
-	
-	"postCreateCommand": "cp /workspaces/auteline/githooks/* .git/hooks/ && gh auth login",
-	
-	"postStartCommand": "echo git username = $(git config user.name) && echo git config email = $(git config user.email)"
+	"features": {
+		"github-cli": "latest",
+		"python": "latest"
+	}
 }
