+++ .vscode/launch.json
@@ -5,17 +5,24 @@
     "version": "0.2.0",
     "configurations": [
         {
-            "name": "Launch",
-            "command": "${workspaceFolder}/bazel-bin/src/main/java/Auteline",
+            "type": "java",
+            "name": "Launch Current File",
             "request": "launch",
-            "type": "node-terminal",
-            "preLaunchTask": "Build Auteline",
+            "mainClass": "${file}"
         },
         {
-            "name": "Test",
-            "command": "${workspaceFolder}/bazel test //...",
+            "type": "java",
+            "name": "Launch ATMTest",
             "request": "launch",
-            "type": "node-terminal",
+            "mainClass": "auteline.ATMTest",
+            "projectName": "app"
+        },
+        {
+            "type": "java",
+            "name": "Launch AutelineApp",
+            "request": "launch",
+            "mainClass": "frame.AutelineApp",
+            "projectName": "app"
         }
     ]
 }
