+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
-    "java.project.referencedLibraries": [
-        "./bazel-bin/external/**/*.jar",
-    ],
+    "gradle.debug": true,
+    "java.import.gradle.java.home": "/usr/lib/jvm/msopenjdk-current/",
+    "java.configuration.updateBuildConfiguration": "interactive"
 }
