+++ app/src/main/java/auteline/DepositSlot.java
@@ -1,0 +1,21 @@
+/**
+ * SMTI06, 54411850, M Haidar Hanif
+ * Task Five: Automated Teller Machine
+ * Auteline | Simple ATM simulator with basic features
+ */
+
+// DepositSlot.java
+// Represents the deposit slot of the ATM
+
+package auteline;
+
+public class DepositSlot {
+
+  // indicates whether envelope was received (always returns true,
+  // because this is only a software simulation of a real deposit slot)
+  public boolean isEnvelopeReceived() {
+    return true; // deposit envelope was received
+  }
+
+}
+
