+++ app/src/main/java/frame/AutelineApp.java
@@ -1,0 +1,20 @@
+/**
+ * SMTI06, 54411850, M Haidar Hanif
+ * Task Five: Automated Teller Machine
+ * Auteline | Simple ATM simulator with basic features
+ */
+
+// AutelineApp.java
+// GUI driver program to test ATM program
+
+package frame;
+
+public class AutelineApp {
+
+  // main method creates and runs the ATM in GUI mode
+  public static void main(String[] args) {
+    UnitFrame realATM = new UnitFrame();
+    realATM.run();
+  }
+
+}
