--- githooks/commit-msg
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash
-
-echo "Hello, world!"
-
-exit 0
