+++ githooks/pre-commit
@@ -1,0 +1,20 @@
+#!/usr/bin/env bash
+
+echo "*****Running unit tests******"
+
+git stash -q --keep-index
+
+gradle test
+
+status=$?
+
+git stash pop -q
+
+echo "*****Done with unit tests******"
+
+if [[ $status -ne 0 ]]
+then
+  echo "Unit Tests Failed!"
+fi
+
+exit $status
