--- scripts/serveTests.sh
@@ -1,3 +1,0 @@
-#! /bin/bash
-
-allure=./allure/bin/allure
