--- src/main/frame/AutelineApp.java
@@ -1,20 +1,0 @@
-/**
- * SMTI06, 54411850, M Haidar Hanif
- * Task Five: Automated Teller Machine
- * Auteline | Simple ATM simulator with basic features
- */
-
-// AutelineApp.java
-// GUI driver program to test ATM program
-
-package main.frame;
-
-public class AutelineApp {
-
-  // main method creates and runs the ATM in GUI mode
-  public static void main(String[] args) {
-    UnitFrame realATM = new UnitFrame();
-    realATM.run();
-  }
-
-}
