--- src/main/frame/BUILD
@@ -1,6 +1,0 @@
-java_binary (
-    name = "AutelineGUI",
-    srcs = glob(["*.java"]),
-    main_class = "main.frame.AutelineApp",
-    deps = ["//src/main/java:AutelineHelpers"],
-)
