--- src/main/java/ATMTest.java
@@ -1,20 +1,0 @@
-/**
- * SMTI06, 54411850, M Haidar Hanif
- * Task Five: Automated Teller Machine
- * Auteline | Simple ATM simulator with basic features
- */
-
-// ATMTest.java
-// Driver program to test ATM program
-
-package main.java;
-
-public class ATMTest {
-
-  // main method creates and runs the ATM
-  public static void main(String[] args) {
-    ATM mobileATM = new ATM();
-    mobileATM.run();
-  }
-
-}
