--- src/main/java/BUILD
@@ -1,14 +1,0 @@
-package(
-    default_visibility = ["//src:__subpackages__"],
-)
-
-java_binary (
-    name = "Auteline",
-    srcs = glob(["*.java"]),
-    main_class = "main.java.ATMTest"
-)
-
-java_library(
-    name = "AutelineHelpers",
-    srcs = glob(["*.java"]),
-)
