--- src/test/java/BUILD
@@ -1,11 +1,0 @@
-
-package(
-    default_visibility = ["//src/test:__subpackages__"],
-)
-
-java_test (
-    name = "TestAccount",
-    srcs = ["TestAccount.java"],
-    test_class = "test.java.TestAccount",
-    deps = [ "//src/main/java:AutelineHelpers", "@maven//:junit_junit" ],
-)
